sa2_text: improved api
Before, there was only an Sa2Text which contained a langage and a set of
text elements. This could be deserialized from a set of bytes. Now
there's a Sa2TextTable which is basically a special form of Strable
which contains Sa2Text elements and a language. Makes working with hint
files more streamlined.